Added fix for issue seen in prepared statement for put into values syntax of column table. Added test for the fix

+++ core/src/main/scala/org/apache/spark/sql/SnappyImplicits.scala
@@ -180,7 +180,7 @@ object snappy extends Serializable {
      *
      * This ignores all SaveMode.
      */
-    def putInto(tableName: String): Int = {
+    def putInto(tableName: String): Long = {
       val df: DataFrame = dfField.get(writer).asInstanceOf[DataFrame]
       val session = df.sparkSession match {
         case sc: SnappySession => sc
